This should be everything needed to extract s3 into a cellarClient and create an IdentityMind cellar client at the application level.
